+++ apps/meteor/client/views/omnichannel/directory/chats/ChatInfo/ChatsContextualBar.tsx
@@ -1,16 +1,16 @@
-import { useRoute, useRouteParameter, useRoomToolbox } from '@rocket.chat/ui-contexts';
-import { useMemo } from 'react';
-import { useTranslation } from 'react-i18next';
-
-import ChatInfo from './ChatInfo';
-import RoomEdit from './RoomEdit';
 import {
 	ContextualbarHeader,
 	ContextualbarIcon,
 	ContextualbarTitle,
 	ContextualbarClose,
 	ContextualbarDialog,
-} from '../../../../../components/Contextualbar';
+} from '@rocket.chat/ui-client';
+import { useRoute, useRouteParameter, useRoomToolbox } from '@rocket.chat/ui-contexts';
+import { useMemo } from 'react';
+import { useTranslation } from 'react-i18next';
+
+import ChatInfo from './ChatInfo';
+import RoomEdit from './RoomEdit';
 import { useRoom } from '../../../../room/contexts/RoomContext';
 
 const PATH = 'live';