+++ apps/meteor/client/views/room/HeaderV2/RoomHeaderE2EESetup.tsx
@@ -12,7 +12,16 @@ const RoomHeaderE2EESetup = ({ room }: RoomHeaderProps) => {
 	const e2eRoomState = useE2EERoomState(room._id);
 
 	if (e2eeState === 'SAVE_PASSWORD' || e2eeState === 'ENTER_PASSWORD' || e2eRoomState === 'WAITING_KEYS') {
-		return <RoomHeader room={room} roomToolbox={<RoomToolboxE2EESetup />} />;
+		return (
+			<RoomHeader
+				room={room}
+				slots={{
+					toolbox: {
+						content: <RoomToolboxE2EESetup />,
+					},
+				}}
+			/>
+		);
 	}
 
 	return <RoomHeader room={room} />;