+++ packages/ui-contexts/src/hooks/useUserCard.ts
@@ -1,0 +1,5 @@
+import { useContext } from 'react';
+
+import { UserCardContext } from '../UserCardContext';
+
+export const useUserCard = () => useContext(UserCardContext);